# api/openapi.json
{
  "components": {
    "schemas": {
      "DecisionOut": {
        "properties": {
          "action": {
            "title": "Action",
            "type": "string"
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "judged": {
            "title": "Judged",
            "type": "boolean"
          },
          "message_id": {
            "title": "Message Id",
            "type": "string"
          },
          "probability": {
            "title": "Probability",
            "type": "number"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "scores": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Scores",
            "type": "object"
          }
        },
        "required": [
          "message_id",
          "action",
          "category",
          "probability",
          "scores",
          "judged",
          "reason"
        ],
        "title": "DecisionOut",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "InMessage": {
        "properties": {
          "author": {
            "default": "",
            "title": "Author",
            "type": "string"
          },
          "author_trusted": {
            "default": false,
            "description": "true skips judgment (moderators, verified staff)",
            "title": "Author Trusted",
            "type": "boolean"
          },
          "channel_topic": {
            "default": "",
            "description": "what the channel/thread is about; used by offtopic",
            "title": "Channel Topic",
            "type": "string"
          },
          "id": {
            "default": "",
            "description": "your id for the message; echoed back",
            "title": "Id",
            "type": "string"
          },
          "text": {
            "maxLength": 8000,
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "text"
        ],
        "title": "InMessage",
        "type": "object"
      },
      "KeyRequest": {
        "properties": {
          "label": {
            "default": "",
            "title": "Label",
            "type": "string"
          },
          "tenant": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Tenant",
            "type": "string"
          }
        },
        "required": [
          "tenant"
        ],
        "title": "KeyRequest",
        "type": "object"
      },
      "ModerateRequest": {
        "properties": {
          "messages": {
            "items": {
              "$ref": "#/components/schemas/InMessage"
            },
            "maxItems": 50,
            "minItems": 1,
            "title": "Messages",
            "type": "array"
          }
        },
        "required": [
          "messages"
        ],
        "title": "ModerateRequest",
        "type": "object"
      },
      "ModerateResponse": {
        "properties": {
          "decisions": {
            "items": {
              "$ref": "#/components/schemas/DecisionOut"
            },
            "title": "Decisions",
            "type": "array"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "usage": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Usage",
            "type": "object"
          }
        },
        "required": [
          "request_id",
          "decisions",
          "usage"
        ],
        "title": "ModerateResponse",
        "type": "object"
      },
      "PolicyIn": {
        "additionalProperties": false,
        "properties": {
          "actions": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actions"
          },
          "rule_actions": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rule Actions"
          },
          "rule_thresholds": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "number"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rule Thresholds"
          },
          "rules": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rules"
          },
          "thresholds": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "number"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thresholds"
          },
          "timeout_minutes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timeout Minutes"
          }
        },
        "title": "PolicyIn",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Moderation decisions for user content. One request per batch to Jev (TypeSafe) returns a probability for each of 9 categories (spam, scam, harassment, nsfw, offtopic, selfharm, doxxing, minors, ai_generated) plus any rules you write in your own words. Thresholds and actions are yours; every decision is logged so you can audit it. ai_generated is experimental and off by default.",
    "title": "jevmod",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/metrics": {
      "get": {
        "operationId": "metrics_metrics_get",
        "responses": {
          "200": {
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Metrics"
      }
    },
    "/v1/decisions": {
      "get": {
        "operationId": "decisions_v1_decisions_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response Decisions V1 Decisions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Decisions"
      }
    },
    "/v1/health": {
      "get": {
        "operationId": "health_v1_health_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Health V1 Health Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Health"
      }
    },
    "/v1/keys": {
      "post": {
        "operationId": "create_key_v1_keys_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KeyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Create Key V1 Keys Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Key"
      }
    },
    "/v1/moderate": {
      "post": {
        "operationId": "moderate_v1_moderate_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModerateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModerateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Moderate"
      }
    },
    "/v1/policy": {
      "get": {
        "operationId": "get_policy_v1_policy_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Policy V1 Policy Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Policy"
      },
      "put": {
        "operationId": "put_policy_v1_policy_put",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PolicyIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Put Policy V1 Policy Put",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Put Policy"
      }
    },
    "/v1/tenant": {
      "delete": {
        "description": "GDPR: forget this tenant's policy, usage and decision log.\n\n`leave_tenant` rather than `delete_tenant`, because this call deletes the subscription row too and the\ncard would otherwise keep being charged for a tenant that no longer exists. It reads the subscription id\nfirst and queues it after, and a deployment with nobody draining that queue - a self-hosted copy, which\nhas no Stripe at all - is left with one harmless row.\n\nThis is the same hole `on_guild_remove` had. The difference here is that nobody is present to read a\nwarning: `/mod forget` is typed by a person who can be told to cancel, and an API call is not.",
        "operationId": "delete_tenant_v1_tenant_delete",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "boolean"
                  },
                  "title": "Response Delete Tenant V1 Tenant Delete",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Tenant"
      }
    }
  }
}